Fix module resolution conflict by moving Game to game/mod.rs
- Move entire Game implementation from core/src/game.rs to core/src/game/mod.rs
- Delete original core/src/game.rs file to resolve module conflict
- Eliminates Rust compilation error: "file for module 'game' found at both locations"
- Game struct and all implementations now properly contained within game module
- No functional changes - purely resolves module organization issue
- All imports remain compatible through module re-exports

Fixes the blocking compilation issue in PR #459
